+++ src/extendedRequest.d.ts
@@ -1,0 +1,12 @@
+import { Session, StoreInterface, SessionOptions } from ".";
+
+declare module "http" {
+  export interface IncomingMessage {
+    sessionId?: string | null;
+    _sessId?: string | null;
+    session: Session;
+    _sessOpts: SessionOptions;
+    _sessStr: string;
+    sessionStore: StoreInterface;
+  }
+}
