--- test/customRequest.d.ts
@@ -1,9 +1,0 @@
-import { Session, StoreInterface } from "../src";
-
-declare module "http" {
-  export interface IncomingMessage {
-    session: Session
-    sessionStore: StoreInterface
-    sessionId: string
-  }
-}
