+++ test/integration/get-initial-props/next-env.d.ts
@@ -1,3 +1,3 @@
 /// <reference types="next" />
 /// <reference types="next/types/global" />
-/// <reference path="../../customRequest.d.ts" />
+/// <reference path="../../../src/extendedRequest.d.ts" />
